pyproject.toml
[tool.poetry]
name = "skeleton"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
aiosqlite = "0.17.0"
anyio = "3.6.1"
asgiref = "3.5.2"
click = "8.1.3"
fastapi = "0.78.0"
h11 = "0.13.0"
httptools = "0.4.0"
idna = "3.3"
iso8601 = "1.0.2"
pydantic = "1.9.1"
pypika-tortoise = "0.1.5"
python-dotenv = "0.20.0"
pytz = "2022.1"
PyYAML = "6.0"
sniffio = "1.2.0"
starlette = "0.19.1"
tortoise-orm = "0.19.1"
typing-extensions = "4.2.0"
uvicorn = "0.17.6"
uvloop = "0.16.0"
watchgod = "0.8.2"
websockets = "10.3"
pycountry = "22.3.5"
SQLAlchemy = "~=1.4.37"
psycopg2-binary = "^2.9.3"
asyncpg = "~=0.25.0"
loguru = "~=0.6.0"
alembic = "~=1.8.0"
boto3 = "~=1.24.17"
cognitojwt = "~=1.4.1"
aiofiles = "^0.8.0"
requests = "^2.28.1"
pySqsListener = "^0.8.10"
SQLAlchemy-Utils = "^0.38.3"
pyreader = "^0.0.1"
fastapi-versioning = "^0.10.0"
passlib = "^1.7.4"
pymongo = "4.3.3"
jwt = "^1.3.1"
jose = "^1.0.0"
crypto = "^1.4.1"
cloudinary = "^1.32.0"
getmac = "^0.9.3"
email-validator = "^2.0.0.post2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
